backend/requirements.txt
flask
flask-cors
Flask-SQLAlchemy
SQLAlchemy
itsdangerous
python-dotenv
Werkzeug
google-genai
google-generativeai
